+++ triangle.go
@@ -2,8 +2,7 @@ package main
 
 // Triangle represents a triangle after rasterization.
 type Triangle struct {
-	points       [3]Vec4
-	texcoords    [3]Tex
-	color        Color
-	averageDepth float64
+	points    [3]Vec4
+	texcoords [3]Tex
+	color     Color
 }